service/requirements.txt
flask>=2.0.2
flask-cors>=3.0.10
python-dotenv>=0.20.0